Move fragments to fragments package
