+++ app/src/main/java/com/cornellappdev/coffee_chats_android/OnboardingActivity.kt
@@ -10,6 +10,10 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentTransaction
+import com.cornellappdev.coffee_chats_android.fragments.EditProfileFragment
+import com.cornellappdev.coffee_chats_android.fragments.PromptsFragment
+import com.cornellappdev.coffee_chats_android.fragments.SocialMediaFragment
+import com.cornellappdev.coffee_chats_android.fragments.UserFieldFragment
 import com.cornellappdev.coffee_chats_android.models.User
 import com.cornellappdev.coffee_chats_android.models.UserField
 import com.cornellappdev.coffee_chats_android.models.UserSession
