+++ app/src/main/java/com/cornellappdev/coffee_chats_android/fragments/AboutFragment.kt
@@ -1,4 +1,4 @@
-package com.cornellappdev.coffee_chats_android
+package com.cornellappdev.coffee_chats_android.fragments
 
 import android.content.Intent
 import android.net.Uri
@@ -7,6 +7,7 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.cornellappdev.coffee_chats_android.R
 import kotlinx.android.synthetic.main.fragment_about.*
 
 class AboutFragment : Fragment() {
