+++ app/src/main/java/com/cornellappdev/coffee_chats_android/fragments/NoMatchFragment.kt
@@ -1,10 +1,11 @@
-package com.cornellappdev.coffee_chats_android
+package com.cornellappdev.coffee_chats_android.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.cornellappdev.coffee_chats_android.R
 
 
 class NoMatchFragment : Fragment() {
