+++ app/src/main/java/com/cornellappdev/coffee_chats_android/fragments/SettingsFragment.kt
@@ -1,12 +1,12 @@
-package com.cornellappdev.coffee_chats_android
+package com.cornellappdev.coffee_chats_android.fragments
 
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Toast
+import com.cornellappdev.coffee_chats_android.ProfileSettingsActivity
+import com.cornellappdev.coffee_chats_android.R
 import kotlinx.android.synthetic.main.fragment_settings.*
 
 class SettingsFragment : Fragment() {
